fix license naming expression
Fix the license name so it fullfills the package.json license section (https://docs.npmjs.com/files/package.json). Compared to 'GPL-V3+' the term 'GPL-3.0-or-later' is also OSI approved.
The benefit from this change would be that the project can be processed in the webjars.com building environment. So it becomes also available as an webjar.